extract HEVC boxes processing into its own file
