+++ libheif/codecs/hevc.h
@@ -31,96 +31,6 @@
 #include <codecs/image_item.h>
 
 
-class Box_hvcC : public Box
-{
-public:
-  Box_hvcC()
-  {
-    set_short_type(fourcc("hvcC"));
-  }
-
-  bool is_essential() const override { return true; }
-
-  struct configuration
-  {
-    uint8_t configuration_version;
-    uint8_t general_profile_space;
-    bool general_tier_flag;
-    uint8_t general_profile_idc;
-    uint32_t general_profile_compatibility_flags;
-
-    static const int NUM_CONSTRAINT_INDICATOR_FLAGS = 48;
-    std::bitset<NUM_CONSTRAINT_INDICATOR_FLAGS> general_constraint_indicator_flags;
-
-    uint8_t general_level_idc;
-
-    uint16_t min_spatial_segmentation_idc;
-    uint8_t parallelism_type;
-    uint8_t chroma_format;
-    uint8_t bit_depth_luma;
-    uint8_t bit_depth_chroma;
-    uint16_t avg_frame_rate;
-
-    uint8_t constant_frame_rate;
-    uint8_t num_temporal_layers;
-    uint8_t temporal_id_nested;
-  };
-
-
-  std::string dump(Indent&) const override;
-
-  bool get_headers(std::vector<uint8_t>* dest) const;
-
-  void set_configuration(const configuration& config) { m_configuration = config; }
-
-  const configuration& get_configuration() const { return m_configuration; }
-
-  void append_nal_data(const std::vector<uint8_t>& nal);
-
-  void append_nal_data(const uint8_t* data, size_t size);
-
-  Error write(StreamWriter& writer) const override;
-
-protected:
-  Error parse(BitstreamRange& range) override;
-
-private:
-  struct NalArray
-  {
-    uint8_t m_array_completeness;
-    uint8_t m_NAL_unit_type;
-
-    std::vector<std::vector<uint8_t> > m_nal_units;
-  };
-
-  configuration m_configuration;
-  uint8_t m_length_size = 4; // default: 4 bytes for NAL unit lengths
-
-  std::vector<NalArray> m_nal_array;
-};
-
-class SEIMessage
-{
-public:
-  virtual ~SEIMessage() = default;
-};
-
-
-class SEIMessage_depth_representation_info : public SEIMessage,
-                                             public heif_depth_representation_info
-{
-public:
-};
-
-
-Error decode_hevc_aux_sei_messages(const std::vector<uint8_t>& data,
-                                   std::vector<std::shared_ptr<SEIMessage>>& msgs);
-
-
-Error parse_sps_for_hvcC_configuration(const uint8_t* sps, size_t size,
-                                       Box_hvcC::configuration* inout_config,
-                                       int* width, int* height);
-
 class ImageItem_HEVC : public ImageItem
 {
 public:
@@ -143,6 +53,9 @@ public:
                                 struct heif_encoder* encoder,
                                 const struct heif_encoding_options& options,
                                 enum heif_image_input_class input_class) override;
+
+  // currently not used
+  void set_preencoded_hevc_image(const std::vector<uint8_t>& data);
 
 protected:
   Result<std::vector<uint8_t>> read_bitstream_configuration_data(heif_item_id itemId) const override;
