+++ libheif/codecs/image_item.cc
@@ -475,130 +475,6 @@ int ImageItem::get_chroma_bits_per_pixel() const
 }
 
 
-#if 0
-int ImageItem::get_luma_bits_per_pixel() const
-{
-  heif_item_id id;
-  Error err = m_heif_context->get_id_of_non_virtual_child_image(m_id, id);
-  if (err) {
-    return -1;
-  }
-
-  // NOLINTNEXTLINE(clang-analyzer-core.CallAndMessage)
-  return m_heif_context->get_heif_file()->get_luma_bits_per_pixel_from_configuration(id);
-}
-
-
-int ImageItem::get_chroma_bits_per_pixel() const
-{
-  heif_item_id id;
-  Error err = m_heif_context->get_id_of_non_virtual_child_image(m_id, id);
-  if (err) {
-    return -1;
-  }
-
-  // NOLINTNEXTLINE(clang-analyzer-core.CallAndMessage)
-  return m_heif_context->get_heif_file()->get_chroma_bits_per_pixel_from_configuration(id);
-}
-#endif
-
-
-void ImageItem::set_preencoded_hevc_image(const std::vector<uint8_t>& data)
-{
-  auto hvcC = std::make_shared<Box_hvcC>();
-
-
-  // --- parse the h265 stream and set hvcC headers and compressed image data
-
-  int state = 0;
-
-  bool first = true;
-  bool eof = false;
-
-  int prev_start_code_start = -1; // init to an invalid value, will always be overwritten before use
-  int start_code_start;
-  int ptr = 0;
-
-  for (;;) {
-    bool dump_nal = false;
-
-    uint8_t c = data[ptr++];
-
-    if (state == 3) {
-      state = 0;
-    }
-
-    if (c == 0 && state <= 1) {
-      state++;
-    }
-    else if (c == 0) {
-      // NOP
-    }
-    else if (c == 1 && state == 2) {
-      start_code_start = ptr - 3;
-      dump_nal = true;
-      state = 3;
-    }
-    else {
-      state = 0;
-    }
-
-    if (ptr == (int) data.size()) {
-      start_code_start = (int) data.size();
-      dump_nal = true;
-      eof = true;
-    }
-
-    if (dump_nal) {
-      if (first) {
-        first = false;
-      }
-      else {
-        std::vector<uint8_t> nal_data;
-        size_t length = start_code_start - (prev_start_code_start + 3);
-
-        nal_data.resize(length);
-
-        assert(prev_start_code_start >= 0);
-        memcpy(nal_data.data(), data.data() + prev_start_code_start + 3, length);
-
-        int nal_type = (nal_data[0] >> 1);
-
-        switch (nal_type) {
-          case 0x20:
-          case 0x21:
-          case 0x22:
-            hvcC->append_nal_data(nal_data);
-            break;
-
-          default: {
-            std::vector<uint8_t> nal_data_with_size;
-            nal_data_with_size.resize(nal_data.size() + 4);
-
-            memcpy(nal_data_with_size.data() + 4, nal_data.data(), nal_data.size());
-            nal_data_with_size[0] = ((nal_data.size() >> 24) & 0xFF);
-            nal_data_with_size[1] = ((nal_data.size() >> 16) & 0xFF);
-            nal_data_with_size[2] = ((nal_data.size() >> 8) & 0xFF);
-            nal_data_with_size[3] = ((nal_data.size() >> 0) & 0xFF);
-
-            m_heif_context->get_heif_file()->append_iloc_data(m_id, nal_data_with_size, 0);
-          }
-            break;
-        }
-      }
-
-      prev_start_code_start = start_code_start;
-    }
-
-    if (eof) {
-      break;
-    }
-  }
-
-  m_heif_context->get_heif_file()->add_property(m_id, hvcC, true);
-}
-
-
 static std::shared_ptr<color_profile_nclx> compute_target_nclx_profile(const std::shared_ptr<HeifPixelImage>& image, const heif_color_profile_nclx* output_nclx_profile)
 {
   auto target_nclx_profile = std::make_shared<color_profile_nclx>();
